tests/wpt/tests/html/semantics/forms/the-textarea-element/textarea-type.html

<!DOCTYPE HTML>
<title>The type IDL attribute</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#dom-textarea-type">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="test">
<textarea></textarea>
</div>
<script>
test(function() {
  assert_equals(document.getElementById("test")
                        .getElementsByTagName("textarea")[0].type,
                "textarea");
}, "Textarea's type attribute should return 'textarea'");
</script>
